+++ src/VaultSharp/V1/AuthMethods/AuthProviderFactory.cs
@@ -6,6 +6,7 @@ using VaultSharp.V1.AuthMethods.Azure;
 using VaultSharp.V1.AuthMethods.Cert;
 using VaultSharp.V1.AuthMethods.Custom;
 using VaultSharp.V1.AuthMethods.GitHub;
+using VaultSharp.V1.AuthMethods.GoogleCloud;
 using VaultSharp.V1.AuthMethods.LDAP;
 using VaultSharp.V1.AuthMethods.Token;
 using VaultSharp.V1.AuthMethods.UserPass;
@@ -34,6 +35,11 @@ namespace VaultSharp.V1.AuthMethods
             if (authInfo.AuthMethodType == AuthMethodType.GitHub)
             {
                 return new GitHubAuthMethodLoginProvider(authInfo as GitHubAuthMethodInfo, polymath);
+            }
+
+            if (authInfo.AuthMethodType == AuthMethodType.GoogleCloud)
+            {
+                return new GoogleCloudAuthMethodLoginProvider(authInfo as GoogleCloudAuthMethodInfo, polymath);
             }
 
             if (authInfo.AuthMethodType == AuthMethodType.LDAP)